Define all of XBeeAddress64 in XBee.h
Since this is a utility class that should really be a very thin wrapper
around the actual data, the compiler should be able to inline all
operations, instead of having to do a full function call just to load or
store a few bytes.

This removes the explicit super-constructor calls and rewrites the
constructors to use initializer lists instead of assigning values in the
body (which is more consise, and allows for constexpr next)

This saves up to a 30 bytes of program spaces on the supplied
examples, probably more for more complex programs.