Introduce constexpr for the XBeeAddress64 constructors
This allows the compiler to better optimize (global) XBeeAdress64
instances.

With the examples sketches, this saves between 44 and 114 bytes of
program space, compiling for the Uno. Even sketches that do not directly
use any addresses shrink, because the initialization of
RemoteAtCommandRequest::broadcastAddress64 can be simplified.

This feature requires C++11 to be enabled. This isn't currently the case
(with Arduino 1.6.4), but is expected to be enabled in a future release.
A preprocessor check is used to ensure that the code compiles with and
without C++11.