Add XBeeWithCallbacks::waitForStatus()
This method loops until a status response is received (e.g. a
AtCommandResponse, TxResponse, etc.) to greatly simplify getting the
result of a command sent. While waiting, any other responses are
processed by the callbacks as normal, so no response gets lost while
waiting.

A matchStatus() helper method is introduced, preparing to reuse it in a
future commit.