Add XBeeWithCallbacks::waitFor()
This allows waiting for an arbitrary API response. Combined with a
user-defined "matching" function, this allows e.g. waiting for a reply
to a packet previously sent in a convenient way.

The main work is done using a waitForInternal() method, so the actual
waitFor() can be a template method (to allow passing a response object
directly, without having to also pass a matching API id). The
implementation could have been put inside waitFor() directly, but then
you would end up with a complete copy of the method for each response
type you wait for.