Allow XBeeWithCallbacks::waitFor() to check status responses too
Now, if a frameId is passed, waitFor() will check for status response
while waiting. If one is received that matches the frameId and has a
non-zero status (e.g. an error), waiting stops and the status is
returned.

When sending a TX packet and subsequently listening for a RX reply, this
allows waiting for a reply and the TX status at the same time and stop
waiting if the TX turns out to have failed. This is particularly
convenient since the TX status can sometimes arrive *before* the RX
reply and sometimes *after*, which makes a sequential waitForStatus()
and waitFor() approach not work.