Add library.properties file
This turns the library into a proper non-legacy library and allows the
Arduino IDE to display some additional info in its library manager. This
also opens up the way to allow installing the library from within the
Arduino IDE, if the library's github url is added to the list of
libraries by the Arduino team.

This fixes the first part of issue #2.