Do not crash if no userdata is passed to the printXXXCb callbacks
Previously, the userdata parameter was used as-is, interpreted as a
Print*. If no userdata was supplied when registering the callback, NULL
is passed, resulting an invalid method lookup and a call to a random
piece of memory (often resulting in a lockup).

By checking for this, forgetting to pass userdata will not lockup, but
also not show any error message (since there is no obvious place to
print it to...). If an invalid pointer is passed, things will still
break, of course.